removed the triplex module since only microrna.org data is used so far
